fra int til float på "Your List"

+++ src/main/java/DataSource/UserMapper.java
@@ -207,7 +207,7 @@ public class UserMapper {
                     String movieTitle = rs.getString("movieTitle");
                     int releaseYear = rs.getInt("movieReleaseYear");
                     String category = rs.getString("movieCategory");
-                    int rating = rs.getInt("movieRating");
+                    float rating = rs.getFloat("movieRating");
                     Movie movie = new Movie(movieTitle, releaseYear, category, rating, movieId);
                     personalMediaList.add(movie);
                 }
@@ -215,7 +215,7 @@ public class UserMapper {
                     String seriesTitle = rs.getString("seriesTitle");
                     int releaseYear = rs.getInt("seriesReleaseYear");
                     String category = rs.getString("seriesCategory");
-                    int rating = rs.getInt("seriesRating");
+                    float rating = rs.getFloat("seriesRating");
                     int season = rs.getInt("season");
                     int episode = rs.getInt("episode");
                     Series series = new Series(seriesTitle, releaseYear, category, rating, seriesId, season, episode);
@@ -225,7 +225,7 @@ public class UserMapper {
                     String audioTitle = rs.getString("audioTitle");
                     int releaseYear = rs.getInt("audioReleaseYear");
                     String category = rs.getString("audioCategory");
-                    int rating = rs.getInt("audioRating");
+                    float rating = rs.getFloat("audioRating");
                     String author = rs.getString("audioAuthor");
                     Audiobooks audiobook = new Audiobooks(audioTitle, releaseYear, category, rating, audioId, author);
                     personalMediaList.add(audiobook);
